organize nightly handler src into src/handlers/externalPropTypeHandler
